requirements.txt
Flask>=2.0.1
scikit-learn>=1.0.2
pandas>=1.3.0
numpy>=1.21.0
joblib>=1.0.1
python-whois>=0.7.3
tld>=0.12.6
requests>=2.26.0
python-dateutil>=2.8.2
pillow>=8.3.1
